+++ examples/DevDing/DiagElement.h
@@ -52,7 +52,7 @@ public:
    * @brief initialize a new Element.
    * @param board The board reference.
    */
-  virtual void init(Board *board);
+  virtual void init(Board *board) override;
 
   /**
    * @brief Set a parameter or property to a new value or start an action.
@@ -61,14 +61,14 @@ public:
    * @return true when property could be changed and the corresponding action
    * could be executed.
    */
-  virtual bool set(const char *name, const char *value);
+  virtual bool set(const char *name, const char *value) override;
 
   /**
    * @brief Activate the Element.
    * @return true when the Element could be activated.
    * @return false when parameters are not usable.
    */
-  virtual void start();
+  virtual void start() override;
 
 private: