+++ src/DigitalOutElement.h
@@ -45,17 +45,17 @@ public:
    * @return true when property could be changed and the corresponding action
    * could be executed.
    */
-  virtual bool set(const char *name, const char *value);
+  virtual bool set(const char *name, const char *value) override;
 
   /**
    * @brief Activate the DigitalOutElement.
    * @return true when activation was good.
    * @return false when activation failed.
    */
-  virtual void start();
+  virtual void start() override;
 
   virtual void pushState(
-      std::function<void(const char *pName, const char *eValue)> callback);
+      std::function<void(const char *pName, const char *eValue)> callback) override;
 
 private:
   /**