+++ src/MenuElement.h
@@ -44,12 +44,12 @@ public:
    * @return true when property could be changed and the corresponding action
    * could be executed.
    */
-  virtual bool set(const char *name, const char *value);
+  virtual bool set(const char *name, const char *value) override;
 
   /**
    * @brief Give some processing time to the timer to check for next action.
    */
-  virtual void loop();
+  virtual void loop() override;
 
 private:
   /** The actual active visualized and changeable value */