+++ src/WeatherFeedElement.h
@@ -53,19 +53,19 @@ public:
    * @return true when property could be changed and the corresponding action
    * could be executed.
    */
-  virtual bool set(const char *name, const char *value);
+  virtual bool set(const char *name, const char *value) override;
 
   /**
    * @brief Activate the WeatherFeedElement.
    * @return true when activation was good.
    * @return false when activation failed.
    */
-  virtual void start();
+  virtual void start() override;
 
   /**
    * @brief check for completed remote actions.
    */
-  virtual void loop();
+  virtual void loop() override;
 
 
   virtual void processHeader(String &key, String &value);